+++ containers/Cloudenv/locales/en.json
@@ -760,5 +760,7 @@
   "cloudenv.create_local_project": "Create a local project",
   "cloudenv.link_local_project": "Link local project",
   "cloudenv.select_local_project": "Select local project",
-  "cloudenv.brand_action_deny": "The current cloud platform does not support this operation"
+  "cloudenv.brand_action_deny": "The current cloud platform does not support this operation",
+  "cloudenv.res_sync": "Resource sync",
+  "cloudenv.bill_sync": "Bill sync"
 }
