+++ src/locales/en.json
@@ -823,6 +823,10 @@
       "failed": "Failed",
       "reject": "Reject"
     },
+    "scheduledtaskBillSync": {
+      "running": "Running",
+      "finished": "Finished"
+    },
     "cloudgroup": {
       "available": "Available",
       "deleting": "Deleting",
